ref(github-gw): remove github-gateway in favor of brigade-github-app (as opt-in chart dependency)
