+++ tests/tests_test.go
@@ -14,7 +14,7 @@ import (
 	"testing"
 
 	"github.com/google/go-github/github"
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 
 	"github.com/Azure/brigade/pkg/storage/kube"
 	"github.com/Azure/brigade/pkg/webhook"
